[1415] Spinta sync improvements: Refactor current implementation
- Checkpoint 2: Move helpers to appropriate dir
